Extract initialization of the REST and /dev/lxd http Server
This is a mechanical move of the REST-API-related setup code in
daemon.go to a new api.go file. It makes the code in the Daemon a bit
more terse and allows for easy re-use of the REST http.Server in tests.
